requirements.txt
Pillow
aiohttp
azure-storage-blob
matplotlib
numpy
opencv-python
pandas
pyyaml
requests
scikit-image
